test/rocksdb: add env variables to skip parts of test
When debugging issues or running local tests, it is
nice to skip rebuilding all of the db_bench tool or
running mkfs on the underlying SPDK device.  So add
SKIP_GIT_CLEAN and SKIP_MKFS checks which allow the
test caller to skip these steps.


Reviewed-on: https://review.gerrithub.io/c/spdk/spdk/+/452261
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ test/blobfs/rocksdb/rocksdb.sh
@@ -40,7 +40,9 @@ timing_enter rocksdb
 timing_enter db_bench_build
 
 pushd $DB_BENCH_DIR
-git clean -x -f -d
+if [ -z "$SKIP_GIT_CLEAN" ]; then
+	git clean -x -f -d
+fi
 $MAKE db_bench $MAKEFLAGS $MAKECONFIG DEBUG_LEVEL=0 SPDK_DIR=$rootdir
 popd
 
@@ -50,9 +52,11 @@ $rootdir/scripts/gen_nvme.sh > $ROCKSDB_CONF
 
 trap 'run_bsdump; rm -f $ROCKSDB_CONF; exit 1' SIGINT SIGTERM EXIT
 
-timing_enter mkfs
-$rootdir/test/blobfs/mkfs/mkfs $ROCKSDB_CONF Nvme0n1
-timing_exit mkfs
+if [ -z "$SKIP_MKFS" ]; then
+	timing_enter mkfs
+	$rootdir/test/blobfs/mkfs/mkfs $ROCKSDB_CONF Nvme0n1
+	timing_exit mkfs
+fi
 
 mkdir -p $output_dir/rocksdb
 RESULTS_DIR=$output_dir/rocksdb
